requirements.txt
requests==2.31.0
pandas==2.1.4
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
ta==0.10.2
